# [IMP] resource_booking: faster tests