Apply lessons from https://youtu.be/lT___hG110Q?t=23285 and disable unnecessary mail stuff.

Before:

    odoo.addons.resource_booking.tests.test_backend tested in 16.18s, 9124 queries
    odoo.addons.resource_booking.tests.test_portal tested in 16.42s, 6685 queries

After:

    odoo.addons.resource_booking.tests.test_backend tested in 13.68s, 7873 queries
    odoo.addons.resource_booking.tests.test_portal tested in 15.17s, 5972 queries

@Tecnativa TT31324